piece out nested types
