+++ sdk/core/Azure.Core/src/Pipeline/HttpClientTransport.cs
@@ -20,7 +20,7 @@ namespace Azure.Core.Pipeline
     /// <summary>
     /// An <see cref="HttpPipelineTransport"/> implementation that uses <see cref="HttpClient"/> as the transport.
     /// </summary>
-    public class HttpClientTransport : HttpPipelineTransport, IDisposable
+    public partial class HttpClientTransport : HttpPipelineTransport, IDisposable
     {
         internal const string MessageForServerCertificateCallback = "MessageForServerCertificateCallback";
 
@@ -359,323 +359,6 @@ namespace Azure.Core.Pipeline
         }
 #endif
 
-        private sealed class PipelineRequest : Request
-        {
-            private ArrayBackedPropertyBag<IgnoreCaseString, object> _headers;
-            private string? _clientRequestId;
-
-            public PipelineRequest()
-            {
-                Method = RequestMethod.Get;
-                _headers = new ArrayBackedPropertyBag<IgnoreCaseString, object>();
-            }
-
-            //public override bool IsHttps => Uri.Scheme == System.Uri.UriSchemeHttps;
-            //public sealed override bool RemoveHeaderValue(string name) => RemoveHeader(name);
-            //public sealed override void SetContent(BinaryData content) => Content = content;
-            //public sealed override void SetHeaderValue(string key, string value) => SetHeader(key, value);
-            //public override bool TryGetHeaderValue(string name, out string? value) => TryGetHeader(name, out value);
-
-            public override string ClientRequestId
-            {
-                get => _clientRequestId ??= Guid.NewGuid().ToString();
-                set
-                {
-                    Argument.AssertNotNull(value, nameof(value));
-                    _clientRequestId = value;
-                }
-            }
-
-            protected internal override void SetHeader(string name, string value)
-            {
-                _headers.Set(new IgnoreCaseString(name), value);
-            }
-
-            protected internal override void AddHeader(string name, string value)
-            {
-                if (_headers.TryAdd(new IgnoreCaseString(name), value, out var existingValue))
-                {
-                    return;
-                }
-
-                switch (existingValue)
-                {
-                    case string stringValue:
-                        _headers.Set(new IgnoreCaseString(name), new List<string> { stringValue, value });
-                        break;
-                    case List<string> listValue:
-                        listValue.Add(value);
-                        break;
-                }
-            }
-
-            protected internal override bool TryGetHeader(string name, [NotNullWhen(true)] out string? value)
-            {
-                if (_headers.TryGetValue(new IgnoreCaseString(name), out var headerValue))
-                {
-                    value = GetHttpHeaderValue(name, headerValue);
-                    return true;
-                }
-
-                value = default;
-                return false;
-            }
-
-            protected internal override bool TryGetHeaderValues(string name, [NotNullWhen(true)] out IEnumerable<string>? values)
-            {
-                if (_headers.TryGetValue(new IgnoreCaseString(name), out var value))
-                {
-                    values = value switch
-                    {
-                        string headerValue => new[] { headerValue },
-                        List<string> headerValues => headerValues,
-                        _ => throw new InvalidOperationException($"Unexpected type for header {name}: {value.GetType()}")
-                    };
-                    return true;
-                }
-
-                values = default;
-                return false;
-            }
-
-            protected internal override bool ContainsHeader(string name) => _headers.TryGetValue(new IgnoreCaseString(name), out _);
-
-            protected internal override bool RemoveHeader(string name) => _headers.TryRemove(new IgnoreCaseString(name));
-
-            protected internal override IEnumerable<HttpHeader> EnumerateHeaders()
-            {
-                for (int i = 0; i < _headers.Count; i++)
-                {
-                    _headers.GetAt(i, out var headerName, out var headerValue);
-                    yield return new HttpHeader(headerName, GetHttpHeaderValue(headerName, headerValue));
-                }
-            }
-
-            public HttpRequestMessage BuildRequestMessage(CancellationToken cancellation)
-            {
-                var method = ToHttpClientMethod(Method);
-                var uri = Uri.ToUri();
-                var currentRequest = new HttpRequestMessage(method, uri);
-                var currentContent = Content != null ? new PipelineContentAdapter(Content, cancellation) : null;
-                currentRequest.Content = currentContent;
-#if NETFRAMEWORK
-                currentRequest.Headers.ExpectContinue = false;
-#endif
-                for (int i = 0; i < _headers.Count; i++)
-                {
-                    _headers.GetAt(i, out var headerName, out var value);
-                    switch (value)
-                    {
-                        case string stringValue:
-                            // Authorization is special cased because it is in the hot path for auth polices that set this header on each request and retry.
-                            if (headerName == HttpHeader.Names.Authorization && AuthenticationHeaderValue.TryParse(stringValue, out var authHeader))
-                            {
-                                currentRequest.Headers.Authorization = authHeader;
-                            }
-                            else if (!currentRequest.Headers.TryAddWithoutValidation(headerName, stringValue))
-                            {
-                                if (currentContent != null && !currentContent.Headers.TryAddWithoutValidation(headerName, stringValue))
-                                {
-                                    throw new InvalidOperationException($"Unable to add header {headerName} to header collection.");
-                                }
-                            }
-                            break;
-                        case List<string> listValue:
-                            if (!currentRequest.Headers.TryAddWithoutValidation(headerName, listValue))
-                            {
-                                if (currentContent != null && !currentContent.Headers.TryAddWithoutValidation(headerName, listValue))
-                                {
-                                    throw new InvalidOperationException($"Unable to add header {headerName} to header collection.");
-                                }
-                            }
-                            break;
-                    }
-                }
-
-                AddPropertiesForBlazor(currentRequest);
-
-                return currentRequest;
-            }
-
-            private static void AddPropertiesForBlazor(HttpRequestMessage currentRequest)
-            {
-                // Disable response caching and enable streaming in Blazor apps
-                // see https://github.com/dotnet/aspnetcore/blob/3143d9550014006080bb0def5b5c96608b025a13/src/Components/WebAssembly/WebAssembly/src/Http/WebAssemblyHttpRequestMessageExtensions.cs
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Create("BROWSER")))
-                {
-                    SetPropertiesOrOptions(currentRequest, "WebAssemblyFetchOptions", new Dictionary<string, object> { { "cache", "no-store" } });
-                    SetPropertiesOrOptions(currentRequest, "WebAssemblyEnableStreamingResponse", true);
-                }
-            }
-
-            private static string GetHttpHeaderValue(string headerName, object value) => value switch
-            {
-                string headerValue => headerValue,
-                List<string> headerValues => string.Join(",", headerValues),
-                _ => throw new InvalidOperationException($"Unexpected type for header {headerName}: {value?.GetType()}")
-            };
-
-            public override void Dispose()
-            {
-                _headers.Dispose();
-                var content = Content;
-                if (content != null)
-                {
-                    Content = null;
-                    content.Dispose();
-                }
-            }
-
-            public override string ToString() => BuildRequestMessage(default).ToString();
-
-            private static readonly HttpMethod s_patch = new HttpMethod("PATCH");
-
-            private static HttpMethod ToHttpClientMethod(RequestMethod requestMethod)
-            {
-                var method = requestMethod.Method;
-
-                // Fast-path common values
-                if (method.Length == 3)
-                {
-                    if (string.Equals(method, "GET", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return HttpMethod.Get;
-                    }
-
-                    if (string.Equals(method, "PUT", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return HttpMethod.Put;
-                    }
-                }
-                else if (method.Length == 4)
-                {
-                    if (string.Equals(method, "POST", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return HttpMethod.Post;
-                    }
-                    if (string.Equals(method, "HEAD", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return HttpMethod.Head;
-                    }
-                }
-                else
-                {
-                    if (string.Equals(method, "PATCH", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return s_patch;
-                    }
-                    if (string.Equals(method, "DELETE", StringComparison.OrdinalIgnoreCase))
-                    {
-                        return HttpMethod.Delete;
-                    }
-                }
-
-                return new HttpMethod(method);
-            }
-
-            private readonly struct IgnoreCaseString : IEquatable<IgnoreCaseString>
-            {
-                private readonly string _value;
-
-                public IgnoreCaseString(string value)
-                {
-                    _value = value;
-                }
-
-                [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                public bool Equals(IgnoreCaseString other) => string.Equals(_value, other._value, StringComparison.OrdinalIgnoreCase);
-                public override bool Equals(object? obj) => obj is IgnoreCaseString other && Equals(other);
-                public override int GetHashCode() => _value.GetHashCode();
-
-                [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                public static bool operator ==(IgnoreCaseString left, IgnoreCaseString right) => left.Equals(right);
-                [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                public static bool operator !=(IgnoreCaseString left, IgnoreCaseString right) => !left.Equals(right);
-                [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                public static implicit operator string(IgnoreCaseString ics) => ics._value;
-            }
-
-            private sealed class PipelineContentAdapter : HttpContent
-            {
-                private readonly RequestContent _pipelineContent;
-                private readonly CancellationToken _cancellationToken;
-
-                public PipelineContentAdapter(RequestContent pipelineContent, CancellationToken cancellationToken)
-                {
-                    _pipelineContent = pipelineContent;
-                    _cancellationToken = cancellationToken;
-                }
-
-                protected override async Task SerializeToStreamAsync(Stream stream, TransportContext? context) => await _pipelineContent.WriteToAsync(stream, _cancellationToken).ConfigureAwait(false);
-
-                protected override bool TryComputeLength(out long length) => _pipelineContent.TryComputeLength(out length);
-
-#if NET5_0_OR_GREATER
-                protected override async Task SerializeToStreamAsync(Stream stream, TransportContext? context, CancellationToken cancellationToken)
-                {
-                    await _pipelineContent!.WriteToAsync(stream, cancellationToken).ConfigureAwait(false);
-                }
-
-                protected override void SerializeToStream(Stream stream, TransportContext? context, CancellationToken cancellationToken)
-                {
-                    _pipelineContent.WriteTo(stream, cancellationToken);
-                }
-#endif
-            }
-        }
-
-        private sealed class HttpClientTransportResponse : Response
-        {
-            private readonly HttpResponseMessage _responseMessage;
-
-            private readonly HttpContent _responseContent;
-
-#pragma warning disable CA2213 // Content stream is intentionally not disposed
-            private Stream? _contentStream;
-#pragma warning restore CA2213
-
-            public HttpClientTransportResponse(string requestId, HttpResponseMessage responseMessage, Stream? contentStream)
-            {
-                ClientRequestId = requestId ?? throw new ArgumentNullException(nameof(requestId));
-                _responseMessage = responseMessage ?? throw new ArgumentNullException(nameof(responseMessage));
-                _contentStream = contentStream;
-                _responseContent = _responseMessage.Content;
-            }
-
-            public override int Status => (int)_responseMessage.StatusCode;
-
-            public override string ReasonPhrase => _responseMessage.ReasonPhrase ?? string.Empty;
-
-            public override Stream? ContentStream
-            {
-                get => _contentStream;
-                set
-                {
-                    // Make sure we don't dispose the content if the stream was replaced
-                    _responseMessage.Content = null;
-
-                    _contentStream = value;
-                }
-            }
-
-            public override string ClientRequestId { get; set; }
-
-            protected internal override bool TryGetHeader(string name, [NotNullWhen(true)] out string? value) => HttpClientTransport.TryGetHeader(_responseMessage.Headers, _responseContent, name, out value);
-
-            protected internal override bool TryGetHeaderValues(string name, [NotNullWhen(true)] out IEnumerable<string>? values) => HttpClientTransport.TryGetHeader(_responseMessage.Headers, _responseContent, name, out values);
-
-            protected internal override bool ContainsHeader(string name) => HttpClientTransport.ContainsHeader(_responseMessage.Headers, _responseContent, name);
-
-            protected internal override IEnumerable<HttpHeader> EnumerateHeaders() => GetHeaders(_responseMessage.Headers, _responseContent);
-
-            public override void Dispose()
-            {
-                _responseMessage?.Dispose();
-                DisposeStreamIfNotBuffered(ref _contentStream);
-            }
-
-            public override string ToString() => _responseMessage.ToString();
-        }
 #if NETCOREAPP
         private static SocketsHttpHandler ApplyOptionsToHandler(SocketsHttpHandler httpHandler, HttpPipelineTransportOptions? options)
         {
